requirements.txt
genanki>=0.13.0
PyYAML>=6.0
Jinja2>=3.0
